Move board to its place; update styles, remove some styled components; cleanup
