+++ src/App.tsx
@@ -1,157 +1,16 @@
-import { DragDropContext, Droppable } from "@hello-pangea/dnd";
-import React, { useState } from "react";
-import styled from "styled-components";
-import { Column } from "./Column";
-
-// export function App() {
-//   return (
-//     <div>
-//       <BrowserRouter>
-//         <ItemsProvider>
-//           <AppRoutes />
-//         </ItemsProvider>
-//       </BrowserRouter>
-//     </div>
-//   );
-// }
-
-const initialData = {
-  tasks: {
-    "task-1": { id: "task-1", content: "Take out the garbage" },
-    "task-2": { id: "task-2", content: "Watch my favorite show" },
-    "task-3": { id: "task-3", content: "Charge my phone" },
-    "task-4": { id: "task-4", content: "Cook dinner", order: "" },
-  },
-  columns: {
-    "column-1": {
-      id: "column-1",
-      title: "To do",
-      taskIds: ["task-1", "task-2", "task-3", "task-4"],
-    },
-    "column-2": {
-      id: "column-2",
-      title: "In progress",
-      taskIds: [],
-    },
-    "column-3": {
-      id: "column-3",
-      title: "Done",
-      taskIds: [],
-    },
-  },
-  columnOrder: ["column-1", "column-2", "column-3"],
-};
-
-const Container = styled.div`
-  display: flex;
-`;
+import React from "react";
+import { BrowserRouter } from "react-router-dom";
+import { AppRoutes } from "./AppRoutes";
+import { ItemsProvider } from "./Items/state/ItemContext";
 
 export function App() {
-  const [state, setState] = useState(initialData);
-
-  const onDragEnd = (result) => {
-    console.log("onDragEnd", result);
-    document.body.style.color = "inherit";
-    const { destination, source, draggableId, type } = result;
-
-    if (!destination) return;
-
-    if (destination.droppableId === source.droppableId && destination.index === source.index)
-      return;
-
-    if (type === "column") {
-      const newColumnOrder = Array.from(state.columnOrder);
-      newColumnOrder.splice(source.index, 1);
-      newColumnOrder.splice(destination.index, 0, draggableId);
-
-      setState((currState) => ({
-        ...currState,
-        columnOrder: newColumnOrder,
-      }));
-      return;
-    }
-
-    const start = state.columns[source.droppableId];
-    const finish = state.columns[destination.droppableId];
-
-    if (start === finish) {
-      const newTaskIds = Array.from(start.taskIds);
-      newTaskIds.splice(source.index, 1);
-      newTaskIds.splice(destination.index, 0, draggableId);
-
-      const newColumn = {
-        ...start,
-        taskIds: newTaskIds,
-      };
-
-      setState((currState) => ({
-        ...currState,
-        columns: {
-          ...state.columns,
-          [newColumn.id]: newColumn,
-        },
-      }));
-    } else {
-      const startTaskIds = Array.from(start.taskIds);
-      startTaskIds.splice(source.index, 1);
-      const newStart = {
-        ...start,
-        taskIds: startTaskIds,
-      };
-
-      const finishTaskIds = Array.from(finish.taskIds);
-      finishTaskIds.splice(destination.index, 0, draggableId);
-      const newFinish = {
-        ...finish,
-        taskIds: finishTaskIds,
-      };
-
-      setState((currState) => ({
-        ...currState,
-        columns: {
-          ...state.columns,
-          [newStart.id]: newStart,
-          [newFinish.id]: newFinish,
-        },
-      }));
-    }
-  };
-
-  const onDragStart = (result) => {
-    console.log("onDragStart", result);
-    document.body.style.color = "orange";
-    document.body.style.transition = "background-color 0.2s ease";
-  };
-  //
-  // const onDragUpdate = (result) => {
-  //   const { destination } = result;
-  //   const opacity = destination ? destination.index / Object.keys(state.tasks).length : 0;
-  //   document.body.style.backgroundColor = `rgba(153, 141, 217, ${opacity})`;
-  // };
-
   return (
-    <DragDropContext
-      onDragStart={onDragStart}
-      // onDragUpdate={onDragUpdate}
-      onDragEnd={onDragEnd}
-    >
-      <Droppable droppableId="all-columns" direction="horizontal" type="column">
-        {(provided) => (
-          <Container
-            {...provided.droppableProps}
-            ref={provided.innerRef}
-            // isDraggingOver={snapshot.isDraggingOver}
-          >
-            {state.columnOrder.map((columnId, index) => {
-              const column = state.columns[columnId];
-              const tasks = column.taskIds.map((taskId) => state.tasks[taskId]);
-
-              return <Column key={columnId} column={column} tasks={tasks} index={index} />;
-            })}
-            {provided.placeholder}
-          </Container>
-        )}
-      </Droppable>
-    </DragDropContext>
+    <div>
+      <BrowserRouter>
+        <ItemsProvider>
+          <AppRoutes />
+        </ItemsProvider>
+      </BrowserRouter>
+    </div>
   );
 }
