--- src/Column.tsx
@@ -1,50 +1,0 @@
-import { Draggable, Droppable } from "@hello-pangea/dnd";
-import React from "react";
-import styled from "styled-components";
-import Task from "./Task";
-
-const Container = styled.div`
-  margin: 8px;
-  border: 1px solid lightgrey;
-  background-color: white;
-  border-radius: 2px;
-  width: 220px;
-  display: flex;
-  flex-direction: column;
-`;
-const Title = styled.h3`
-  padding: 8px;
-`;
-const TaskList = styled.div`
-  padding: 8px;
-  transition: background-color 0.2s ease;
-  background-color: ${(props) => (props.isDraggingOver ? "skyblue" : "inherit")};
-  flex-grow: 1;
-  min-height: 100px;
-`;
-
-export function Column({ column, index, tasks }: { tasks: any; column: any; index: number }) {
-  return (
-    <Draggable draggableId={column.id} index={index}>
-      {(provided) => (
-        <Container {...provided.draggableProps} ref={provided.innerRef}>
-          <Title {...provided.dragHandleProps}>{column.title}</Title>
-          <Droppable droppableId={column.id}>
-            {(provided, snapshot) => (
-              <TaskList
-                {...provided.droppableProps}
-                ref={provided.innerRef}
-                isDraggingOver={snapshot.isDraggingOver}
-              >
-                {tasks.map((task, index) => (
-                  <Task key={task.id} task={task} index={index} />
-                ))}
-                {provided.placeholder}
-              </TaskList>
-            )}
-          </Droppable>
-        </Container>
-      )}
-    </Draggable>
-  );
-}
