--- src/Task.tsx
@@ -1,28 +1,0 @@
-import { Draggable } from "@hello-pangea/dnd";
-import React from "react";
-import styled from "styled-components";
-
-const Container = styled.div`
-  padding: 8px;
-  border: 1px solid lightgrey;
-  border-radius: 2px;
-  margin-bottom: 8px;
-  background-color: ${(props) => (props.isDragging ? "lightgreen" : "white")};
-`;
-
-export default function Task({ index, task }: { index: number; task: any }) {
-  return (
-    <Draggable draggableId={task.id} index={index}>
-      {(provided, snapshot) => (
-        <Container
-          {...provided.draggableProps}
-          {...provided.dragHandleProps}
-          ref={provided.innerRef}
-          isDragging={snapshot.isDragging}
-        >
-          {task.content}
-        </Container>
-      )}
-    </Draggable>
-  );
-}
